+++ mirrord-layer/src/macros.rs
@@ -89,5 +89,20 @@ macro_rules! hook_symbol {
     };
 }
 
+#[macro_export]
+macro_rules! graceful_exit {
+    ($($arg:tt)+) => {
+        eprintln!($($arg)+);
+        graceful_exit!()
+    };
+    () => {
+        nix::sys::signal::kill(
+            nix::unistd::Pid::from_raw(std::process::id() as i32),
+            nix::sys::signal::Signal::SIGTERM,
+        )
+        .expect("unable to graceful exit")
+    };
+}
+
 #[cfg(target_os = "linux")]
 pub(crate) use hook_symbol;